+++ src/main/java/demo/rest/DemoApplication.java
@@ -1,5 +1,5 @@
 
-package com.example.demo;
+package demo.rest;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.GetMapping;
