+++ src/test/java/demo/rest/DemoApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package demo.rest;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
